Create PyModule-xml.etree.ElementTree.txt
xml.etree.ElementTree (from now on, ET in short): It provides a lightweight Pythonic API, backed by an efficient C implementation, 
for parsing and creating XML. Compared to DOM, ET is much faster [3] and has a more pleasant API to work with. Compared to SAX, 
there is ET.iterparse which also provides "on the fly" parsing without loading the whole document into memory. The performance is 
on par with SAX, but the API is higher level and much more convenient to use; it will be demonstrated later in the article.

xml.dom.* modules : implement the W3C DOM API. If you're used to working with the DOM API or have some requirement to do so, this 
package can help you. Note that there are several modules in the xml.dom package, representing different tradeoffs between performance 
and expressivity.

xml.sax.* modules : implement the SAX API, which trades convenience for speed and memory consumption. SAX is an event-based API 
meant to parse huge documents "on the fly" without loading them wholly into memory As opposed to DOM, which loads the whole document 
into memory and allows "random access" to its elements at any depth..

It is recommended to always use ET for XML processing in Python, unless you have very specific needs that may call for the other 
solutions.